+++ src/layers.py
@@ -8,368 +8,9 @@ import functools
 import torch
 from torch import nn
 import torch.nn.functional as F
-from torch.distributions import Categorical
 import torch_geometric
 from torch_geometric.nn import MessagePassing
 from torch_scatter import scatter_add
-
-#########################################################################
-
-class NonAutoregressiveMultiGNN(torch.nn.Module):
-    '''
-    Non-Autoregressive GVP-GNN for **multiple** structure-conditioned RNA design.
-    
-    Takes in RNA structure graphs of type `torch_geometric.data.Data` 
-    or `torch_geometric.data.Batch` and returns a categorical distribution
-    over 4 bases at each position in a `torch.Tensor` of shape [n_nodes, 4].
-    
-    The standard forward pass requires sequence information as input
-    and should be used for training or evaluating likelihood.
-    For sampling or design, use `self.sample`.
-    
-    :param node_in_dim: node dimensions in input graph
-    :param node_h_dim: node dimensions to use in GVP-GNN layers
-    :param node_in_dim: edge dimensions in input graph
-    :param edge_h_dim: edge dimensions to embed in GVP-GNN layers
-    :param num_layers: number of GVP-GNN layers in encoder/decoder
-    :param drop_rate: rate to use in all dropout layers
-    :param out_dim: output dimension (4 bases)
-    '''
-    def __init__(
-        self,
-        node_in_dim = (64, 4), 
-        node_h_dim = (128, 16), 
-        edge_in_dim = (32, 1), 
-        edge_h_dim = (32, 1),
-        num_layers = 3, 
-        drop_rate = 0.1,
-        out_dim = 4,
-    ):
-        super().__init__()
-        self.node_in_dim = node_in_dim
-        self.node_h_dim = node_h_dim
-        self.edge_in_dim = edge_in_dim
-        self.edge_h_dim = edge_h_dim
-        self.num_layers = num_layers
-        self.out_dim = out_dim
-        activations = (F.relu, None)
-        
-        # Node input embedding
-        self.W_v = torch.nn.Sequential(
-            LayerNorm(self.node_in_dim),
-            GVP(self.node_in_dim, self.node_h_dim,
-                activations=(None, None), vector_gate=True)
-        )
-
-        # Edge input embedding
-        self.W_e = torch.nn.Sequential(
-            LayerNorm(self.edge_in_dim),
-            GVP(self.edge_in_dim, self.edge_h_dim, 
-                activations=(None, None), vector_gate=True)
-        )
-        
-        # Encoder layers (supports multiple conformations)
-        self.encoder_layers = nn.ModuleList(
-                MultiGVPConvLayer(self.node_h_dim, self.edge_h_dim, 
-                                  activations=activations, vector_gate=True,
-                                  drop_rate=drop_rate)
-            for _ in range(num_layers))
-        
-        # Output
-        self.W_out = torch.nn.Sequential(
-            LayerNorm(self.node_h_dim),
-            GVP(self.node_h_dim, self.node_h_dim,
-                activations=(None, None), vector_gate=True),
-            GVP(self.node_h_dim, (self.out_dim, 0), 
-                activations=(None, None))   
-        )
-    
-    def forward(self, batch):
-
-        h_V = (batch.node_s, batch.node_v)
-        h_E = (batch.edge_s, batch.edge_v)
-        edge_index = batch.edge_index
-        
-        h_V = self.W_v(h_V)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-        h_E = self.W_e(h_E)  # (n_edges, n_conf, d_se), (n_edges, n_conf, d_ve, 3)
-
-        for layer in self.encoder_layers:
-            h_V = layer(h_V, edge_index, h_E)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-
-        # Pool multi-conformation features: 
-        # nodes: (n_nodes, d_s), (n_nodes, d_v, 3)
-        # edges: (n_edges, d_se), (n_edges, d_ve, 3)
-        h_V, h_E = self.pool_multi_conf(h_V, h_E, batch.mask_confs, edge_index)
-
-        logits = self.W_out(h_V)  # (n_nodes, out_dim)
-        
-        return logits
-    
-    def sample(self, batch, n_samples, temperature=0.1):
-        '''
-        Samples sequences from the distribution learned by the model.
-        
-        :param batch: mini-batch (only supports one sample at a time)
-        :param n_samples: number of samples
-        :param temperature: temperature to use in softmax 
-                            over the categorical distribution
-        
-        :return: int `torch.Tensor` of shape [n_samples, n_nodes] based on the
-                 residue-to-int mapping of the original training data
-        '''
-        
-        with torch.no_grad():
-
-            h_V = (batch.node_s, batch.node_v)
-            h_E = (batch.edge_s, batch.edge_v)
-            edge_index = batch.edge_index
-        
-            h_V = self.W_v(h_V)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-            h_E = self.W_e(h_E)  # (n_edges, n_conf, d_se), (n_edges, n_conf, d_ve, 3)
-            
-            for layer in self.encoder_layers:
-                h_V = layer(h_V, edge_index, h_E)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-            
-            # Pool multi-conformation features
-            h_V, h_E = self.pool_multi_conf(h_V, h_E, batch.mask_confs, edge_index)
-            
-            logits = self.W_out(h_V)  # (n_nodes, out_dim)
-
-            logits /= temperature
-            probs = F.softmax(logits, dim=-1)
-
-            seq = torch.multinomial(probs, n_samples, replacement=True)  # (n_nodes, n_samples)
-                
-            return seq.permute(1, 0)
-        
-    def pool_multi_conf(self, h_V, h_E, mask_confs, edge_index):
-        
-        # True num_conf for masked mean pooling
-        n_conf_true = mask_confs.sum(1, keepdim=True)  # (n_nodes, 1)
-        
-        # Mask scalar features
-        mask = mask_confs.unsqueeze(2)  # (n_nodes, n_conf, 1)
-        h_V0 = h_V[0] * mask
-        h_E0 = h_E[0] * mask[edge_index[0]]
-
-        # Mask vector features
-        mask = mask.unsqueeze(3)  # (n_nodes, n_conf, 1, 1)
-        h_V1 = h_V[1] * mask
-        h_E1 = h_E[1] * mask[edge_index[0]]
-        
-        # Average pooling multi-conformation features
-        h_V = (h_V0.sum(dim=1) / n_conf_true,               # (n_nodes, d_s)
-               h_V1.sum(dim=1) / n_conf_true.unsqueeze(2))  # (n_nodes, d_v, 3)
-        h_E = (h_E0.sum(dim=1) / n_conf_true[edge_index[0]],               # (n_edges, d_se)
-               h_E1.sum(dim=1) / n_conf_true[edge_index[0]].unsqueeze(2))  # (n_edges, d_ve, 3)
-
-        return h_V, h_E
-
-#########################################################################
-
-class AutoregressiveMultiGNN(torch.nn.Module):
-    '''
-    Autoregressive GVP-GNN for **multiple** structure-conditioned RNA design.
-    
-    Takes in RNA structure graphs of type `torch_geometric.data.Data` 
-    or `torch_geometric.data.Batch` and returns a categorical distribution
-    over 4 bases at each position in a `torch.Tensor` of shape [n_nodes, 4].
-    
-    The standard forward pass requires sequence information as input
-    and should be used for training or evaluating likelihood.
-    For sampling or design, use `self.sample`.
-    
-    :param node_in_dim: node dimensions in input graph
-    :param node_h_dim: node dimensions to use in GVP-GNN layers
-    :param node_in_dim: edge dimensions in input graph
-    :param edge_h_dim: edge dimensions to embed in GVP-GNN layers
-    :param num_layers: number of GVP-GNN layers in encoder/decoder
-    :param drop_rate: rate to use in all dropout layers
-    :param out_dim: output dimension (4 bases)
-    '''
-    def __init__(
-        self,
-        node_in_dim = (64, 4), 
-        node_h_dim = (128, 16), 
-        edge_in_dim = (32, 1), 
-        edge_h_dim = (32, 1),
-        num_layers = 3, 
-        drop_rate = 0.1,
-        out_dim = 4,
-    ):
-        super().__init__()
-        self.node_in_dim = node_in_dim
-        self.node_h_dim = node_h_dim
-        self.edge_in_dim = edge_in_dim
-        self.edge_h_dim = edge_h_dim
-        self.num_layers = num_layers
-        self.out_dim = out_dim
-        activations = (F.relu, None)
-        
-        # Node input embedding
-        self.W_v = torch.nn.Sequential(
-            LayerNorm(self.node_in_dim),
-            GVP(self.node_in_dim, self.node_h_dim,
-                activations=(None, None), vector_gate=True)
-        )
-
-        # Edge input embedding
-        self.W_e = torch.nn.Sequential(
-            LayerNorm(self.edge_in_dim),
-            GVP(self.edge_in_dim, self.edge_h_dim, 
-                activations=(None, None), vector_gate=True)
-        )
-        
-        # Encoder layers (supports multiple conformations)
-        self.encoder_layers = nn.ModuleList(
-                MultiGVPConvLayer(self.node_h_dim, self.edge_h_dim, 
-                                  activations=activations, vector_gate=True,
-                                  drop_rate=drop_rate)
-            for _ in range(num_layers))
-        
-        # Decoder layers
-        self.W_s = nn.Embedding(self.out_dim, self.out_dim)
-        self.edge_h_dim = (self.edge_h_dim[0] + self.out_dim, self.edge_h_dim[1])
-        self.decoder_layers = nn.ModuleList(
-                GVPConvLayer(self.node_h_dim, self.edge_h_dim,
-                             activations=activations, vector_gate=True, 
-                             drop_rate=drop_rate, autoregressive=True) 
-            for _ in range(num_layers))
-        
-        # Output
-        self.W_out = GVP(self.node_h_dim, (self.out_dim, 0), activations=(None, None))
-    
-    def forward(self, batch):
-
-        h_V = (batch.node_s, batch.node_v)
-        h_E = (batch.edge_s, batch.edge_v)
-        edge_index = batch.edge_index
-        seq = batch.seq
-
-        h_V = self.W_v(h_V)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-        h_E = self.W_e(h_E)  # (n_edges, n_conf, d_se), (n_edges, n_conf, d_ve, 3)
-
-        for layer in self.encoder_layers:
-            h_V = layer(h_V, edge_index, h_E)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-
-        # Pool multi-conformation features: 
-        # nodes: (n_nodes, d_s), (n_nodes, d_v, 3)
-        # edges: (n_edges, d_se), (n_edges, d_ve, 3)
-        h_V, h_E = self.pool_multi_conf(h_V, h_E, batch.mask_confs, edge_index)
-
-        encoder_embeddings = h_V
-        
-        h_S = self.W_s(seq)
-        h_S = h_S[edge_index[0]]
-        h_S[edge_index[0] >= edge_index[1]] = 0
-        h_E = (torch.cat([h_E[0], h_S], dim=-1), h_E[1])
-        
-        for layer in self.decoder_layers:
-            h_V = layer(h_V, edge_index, h_E, autoregressive_x = encoder_embeddings)
-        
-        logits = self.W_out(h_V)
-        
-        return logits
-    
-    def sample(self, batch, n_samples, temperature=0.1):
-        '''
-        Samples sequences autoregressively from the distribution
-        learned by the model.
-        
-        :param batch: mini-batch (only supports one sample at a time)
-        :param n_samples: number of samples
-        :param temperature: temperature to use in softmax 
-                            over the categorical distribution
-        
-        :return: int `torch.Tensor` of shape [n_samples, n_nodes] based on the
-                 residue-to-int mapping of the original training data
-        '''
-        
-        with torch.no_grad():
-
-            h_V = (batch.node_s, batch.node_v)
-            h_E = (batch.edge_s, batch.edge_v)
-            edge_index = batch.edge_index
-        
-            device = edge_index.device
-            num_nodes = h_V[0].shape[0]
-            
-            h_V = self.W_v(h_V)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-            h_E = self.W_e(h_E)  # (n_edges, n_conf, d_se), (n_edges, n_conf, d_ve, 3)
-            
-            for layer in self.encoder_layers:
-                h_V = layer(h_V, edge_index, h_E)  # (n_nodes, n_conf, d_s), (n_nodes, n_conf, d_v, 3)
-            
-            # Pool multi-conformation features
-            h_V, h_E = self.pool_multi_conf(h_V, h_E, batch.mask_confs, edge_index)
-            
-            # Repeat features for sampling n_samples times
-            h_V = (h_V[0].repeat(n_samples, 1),
-                   h_V[1].repeat(n_samples, 1, 1))
-            h_E = (h_E[0].repeat(n_samples, 1),
-                   h_E[1].repeat(n_samples, 1, 1))
-            
-            # Expand edge index for autoregressive decoding
-            edge_index = edge_index.expand(n_samples, -1, -1)
-            offset = num_nodes * torch.arange(n_samples, device=device).view(-1, 1, 1)
-            edge_index = torch.cat(tuple(edge_index + offset), dim=-1)
-            
-            seq = torch.zeros(n_samples * num_nodes, device=device, dtype=torch.int)
-            h_S = torch.zeros(n_samples * num_nodes, self.out_dim, device=device)
-    
-            h_V_cache = [(h_V[0].clone(), h_V[1].clone()) for _ in self.decoder_layers]
-            
-            # Decode one token at a time
-            for i in range(num_nodes):
-                
-                h_S_ = h_S[edge_index[0]]
-                h_S_[edge_index[0] >= edge_index[1]] = 0
-                h_E_ = (torch.cat([h_E[0], h_S_], dim=-1), h_E[1])
-                        
-                edge_mask = edge_index[1] % num_nodes == i
-                edge_index_ = edge_index[:, edge_mask]
-                h_E_ = tuple_index(h_E_, edge_mask)
-                node_mask = torch.zeros(n_samples * num_nodes, device=device, dtype=torch.bool)
-                node_mask[i::num_nodes] = True
-                
-                for j, layer in enumerate(self.decoder_layers):
-                    out = layer(h_V_cache[j], edge_index_, h_E_,
-                               autoregressive_x=h_V_cache[0], node_mask=node_mask)
-                    
-                    out = tuple_index(out, node_mask)
-                    
-                    if j < len(self.decoder_layers)-1:
-                        h_V_cache[j+1][0][i::num_nodes] = out[0]
-                        h_V_cache[j+1][1][i::num_nodes] = out[1]
-                    
-                logits = self.W_out(out)
-                seq[i::num_nodes] = Categorical(logits=logits / temperature).sample()
-                h_S[i::num_nodes] = self.W_s(seq[i::num_nodes])
-                
-            return seq.view(n_samples, num_nodes)
-        
-    def pool_multi_conf(self, h_V, h_E, mask_confs, edge_index):
-        
-        # True num_conf for masked mean pooling
-        n_conf_true = mask_confs.sum(1, keepdim=True)  # (n_nodes, 1)
-        
-        # Mask scalar features
-        mask = mask_confs.unsqueeze(2)  # (n_nodes, n_conf, 1)
-        h_V0 = h_V[0] * mask
-        h_E0 = h_E[0] * mask[edge_index[0]]
-
-        # Mask vector features
-        mask = mask.unsqueeze(3)  # (n_nodes, n_conf, 1, 1)
-        h_V1 = h_V[1] * mask
-        h_E1 = h_E[1] * mask[edge_index[0]]
-        
-        # Average pooling multi-conformation features
-        h_V = (h_V0.sum(dim=1) / n_conf_true,               # (n_nodes, d_s)
-               h_V1.sum(dim=1) / n_conf_true.unsqueeze(2))  # (n_nodes, d_v, 3)
-        h_E = (h_E0.sum(dim=1) / n_conf_true[edge_index[0]],               # (n_edges, d_se)
-               h_E1.sum(dim=1) / n_conf_true[edge_index[0]].unsqueeze(2))  # (n_edges, d_ve, 3)
-
-        return h_V, h_E
 
 #########################################################################
 
@@ -402,9 +43,10 @@ class GVPConvLayer(nn.Module):
             n_feedforward=2, 
             drop_rate=.1,
             autoregressive=False, 
-            activations=(F.relu, torch.sigmoid), 
+            activations=(F.silu, torch.sigmoid), 
             vector_gate=True,
-            residual=True
+            residual=True,
+            norm_first=False,
         ):
         
         super(GVPConvLayer, self).__init__()
@@ -418,7 +60,7 @@ class GVPConvLayer(nn.Module):
 
         ff_func = []
         if n_feedforward == 1:
-            ff_func.append(GVP_(node_dims, node_dims, activations=(None, None)))
+            ff_func.append(GVP_(node_dims, node_dims))
         else:
             hid_dims = 4*node_dims[0], 2*node_dims[1]
             ff_func.append(GVP_(node_dims, hid_dims))
@@ -427,6 +69,7 @@ class GVPConvLayer(nn.Module):
             ff_func.append(GVP_(hid_dims, node_dims, activations=(None, None)))
         self.ff_func = nn.Sequential(*ff_func)
         self.residual = residual
+        self.norm_first = norm_first
 
     def forward(self, x, edge_index, edge_attr,
                 autoregressive_x=None, node_mask=None):
@@ -436,7 +79,7 @@ class GVPConvLayer(nn.Module):
         :param edge_attr: tuple (s, V) of `torch.Tensor`
         :param autoregressive_x: tuple (s, V) of `torch.Tensor`. 
                 If not `None`, will be used as src node embeddings
-                for forming messages where src >= dst. The corrent node 
+                for forming messages where src >= dst. The current node 
                 embeddings `x` will still be the base of the update and the 
                 pointwise feedforward.
         :param node_mask: array of type `bool` to index into the first
@@ -463,16 +106,23 @@ class GVPConvLayer(nn.Module):
             dh = dh[0] / count, dh[1] / count.unsqueeze(-1)
 
         else:
-            dh = self.conv(x, edge_index, edge_attr)
+            if self.norm_first:
+                dh = self.conv(self.norm[0](x), edge_index, edge_attr)
+            else:
+                dh = self.conv(x, edge_index, edge_attr)
         
         if node_mask is not None:
             x_ = x
             x, dh = tuple_index(x, node_mask), tuple_index(dh, node_mask)
-            
-        x = self.norm[0](tuple_sum(x, self.dropout[0](dh))) if self.residual else dh
         
-        dh = self.ff_func(x)
-        x = self.norm[1](tuple_sum(x, self.dropout[1](dh))) if self.residual else dh
+        if self.norm_first:
+            x = tuple_sum(x, self.dropout[0](dh))
+            dh = self.ff_func(self.norm[1](x))
+            x = tuple_sum(x, self.dropout[1](dh))
+        else:
+            x = self.norm[0](tuple_sum(x, self.dropout[0](dh))) if self.residual else dh
+            dh = self.ff_func(x)
+            x = self.norm[1](tuple_sum(x, self.dropout[1](dh))) if self.residual else dh
         
         if node_mask is not None:
             x_[0][node_mask], x_[1][node_mask] = x[0], x[1]
@@ -501,7 +151,7 @@ class GVPConv(MessagePassing):
     '''
     def __init__(self, in_dims, out_dims, edge_dims,
                  n_layers=3, module_list=None, aggr="mean", 
-                 activations=(F.relu, torch.sigmoid), vector_gate=True):
+                 activations=(F.silu, torch.sigmoid), vector_gate=True):
         super(GVPConv, self).__init__(aggr=aggr)
         self.si, self.vi = in_dims
         self.so, self.vo = out_dims
@@ -515,7 +165,7 @@ class GVPConv(MessagePassing):
             if n_layers == 1:
                 module_list.append(
                     GVP_((2*self.si + self.se, 2*self.vi + self.ve), 
-                        (self.so, self.vo), activations=(None, None)))
+                        (self.so, self.vo)))
             else:
                 module_list.append(
                     GVP_((2*self.si + self.se, 2*self.vi + self.ve), out_dims)
@@ -558,9 +208,10 @@ class MultiGVPConvLayer(nn.Module):
             n_message=3, 
             n_feedforward=2, 
             drop_rate=.1,
-            activations=(F.relu, torch.sigmoid), 
+            activations=(F.silu, torch.sigmoid), 
             vector_gate=True,
-            residual=True
+            residual=True,
+            norm_first=False,
         ):
         super(MultiGVPConvLayer, self).__init__()
         self.conv = MultiGVPConv(node_dims, node_dims, edge_dims, n_message,
@@ -572,7 +223,7 @@ class MultiGVPConvLayer(nn.Module):
 
         ff_func = []
         if n_feedforward == 1:
-            ff_func.append(GVP_(node_dims, node_dims, activations=(None, None)))
+            ff_func.append(GVP_(node_dims, node_dims))
         else:
             hid_dims = 4*node_dims[0], 2*node_dims[1]
             ff_func.append(GVP_(node_dims, hid_dims))
@@ -581,6 +232,7 @@ class MultiGVPConvLayer(nn.Module):
             ff_func.append(GVP_(hid_dims, node_dims, activations=(None, None)))
         self.ff_func = nn.Sequential(*ff_func)
         self.residual = residual
+        self.norm_first = norm_first
 
     def forward(self, x, edge_index, edge_attr):
         '''
@@ -588,10 +240,56 @@ class MultiGVPConvLayer(nn.Module):
         :param edge_index: array of shape [2, n_edges]
         :param edge_attr: tuple (s, V) of `torch.Tensor`
         '''
-        dh = self.conv(x, edge_index, edge_attr)    
-        x = self.norm[0](tuple_sum(x, self.dropout[0](dh))) if self.residual else dh
-        dh = self.ff_func(x)
-        x = self.norm[1](tuple_sum(x, self.dropout[1](dh))) if self.residual else dh
+        if self.norm_first:
+            dh = self.conv(self.norm[0](x), edge_index, edge_attr)
+            x = tuple_sum(x, self.dropout[0](dh))
+            dh = self.ff_func(self.norm[1](x))
+            x = tuple_sum(x, self.dropout[1](dh))
+        else:
+            dh = self.conv(x, edge_index, edge_attr)
+            x = self.norm[0](tuple_sum(x, self.dropout[0](dh))) if self.residual else dh
+            dh = self.ff_func(x)
+            x = self.norm[1](tuple_sum(x, self.dropout[1](dh))) if self.residual else dh
+        return x
+
+class NoFFMultiGVPConvLayer(nn.Module):
+    '''
+    GVPConvLayer for handling multiple conformations (encoder-only)
+
+    No position-wise feedforward module.
+    '''
+    def __init__(
+            self, 
+            node_dims, 
+            edge_dims,
+            n_message=2, 
+            drop_rate=.1,
+            activations=(F.silu, torch.sigmoid), 
+            vector_gate=True,
+            residual=True,
+            norm_first=True,
+        ):
+        super(NoFFMultiGVPConvLayer, self).__init__()
+        self.conv = MultiGVPConv(node_dims, node_dims, edge_dims, n_message,
+                                 aggr="mean", activations=activations, vector_gate=vector_gate)
+        self.norm = LayerNorm(node_dims)
+        self.dropout = Dropout(drop_rate)
+
+        self.residual = residual
+        self.norm_first = norm_first
+
+    def forward(self, x, edge_index, edge_attr):
+        '''
+        :param x: tuple (s, V) of `torch.Tensor`
+        :param edge_index: array of shape [2, n_edges]
+        :param edge_attr: tuple (s, V) of `torch.Tensor`
+        '''
+        if self.norm_first:
+            dh = self.conv(self.norm(x), edge_index, edge_attr)
+            x = tuple_sum(x, self.dropout(dh))
+        else:
+            dh = self.conv(x, edge_index, edge_attr)
+            x = self.norm(tuple_sum(x, self.dropout(dh))) if self.residual else dh
         return x
 
 class MultiGVPConv(MessagePassing):
@@ -600,7 +298,7 @@ class MultiGVPConv(MessagePassing):
     '''
     def __init__(self, in_dims, out_dims, edge_dims,
                  n_layers=3, module_list=None, aggr="mean", 
-                 activations=(F.relu, torch.sigmoid), vector_gate=True):
+                 activations=(F.silu, torch.sigmoid), vector_gate=True):
         super(MultiGVPConv, self).__init__(aggr=aggr)
         self.si, self.vi = in_dims
         self.so, self.vo = out_dims
@@ -614,7 +312,7 @@ class MultiGVPConv(MessagePassing):
             if n_layers == 1:
                 module_list.append(
                     GVP_((2*self.si + self.se, 2*self.vi + self.ve), 
-                        (self.so, self.vo), activations=(None, None)))
+                        (self.so, self.vo)))
             else:
                 module_list.append(
                     GVP_((2*self.si + self.se, 2*self.vi + self.ve), out_dims)
@@ -670,7 +368,7 @@ class GVP(nn.Module):
                         (vector_act will be used as sigma^+ in vector gating if `True`)
     '''
     def __init__(self, in_dims, out_dims, h_dim=None,
-                 activations=(F.relu, torch.sigmoid), vector_gate=True):
+                 activations=(F.silu, torch.sigmoid), vector_gate=True):
         super(GVP, self).__init__()
         self.si, self.vi = in_dims
         self.so, self.vo = out_dims
